refactor: 기존에 존재하던 파일의 패키지 위치 변경
