+++ src/main/java/io/hhplus/tdd/common/advice/ApiControllerAdvice.java
@@ -1,5 +1,6 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.common.advice;
 
+import io.hhplus.tdd.common.model.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
