+++ src/main/java/io/hhplus/tdd/common/model/ErrorResponse.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.common.model;
 
 public record ErrorResponse(
         String code,
