+++ src/main/java/io/hhplus/tdd/point/domain/model/entity/PointHistory.java
@@ -1,4 +1,6 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.domain.model.entity;
+
+import io.hhplus.tdd.point.domain.model.vo.TransactionType;
 
 public record PointHistory(
         long id,
