+++ src/main/java/io/hhplus/tdd/point/presentation/PointController.java
@@ -1,5 +1,7 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.presentation;
 
+import io.hhplus.tdd.point.domain.model.entity.PointHistory;
+import io.hhplus.tdd.point.domain.model.entity.UserPoint;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.*;
